requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
openai==0.27.8
pydub==0.25.1
pyTelegramBotAPI==4.12.0
python-decouple==3.8
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
telebot==0.0.5
tqdm==4.65.0
urllib3==2.0.4
yarl==1.9.2
